move fetch to new datasets package
