+++ django_echarts/datasets/fetch.py
