+++ setup.py
@@ -25,6 +25,7 @@ lib_classifiers = [
 
 packages = [
     'django_echarts',
+    'django_echarts.datasets',
     'django_echarts.management',
     'django_echarts.management.commands',
     'django_echarts.plugins',
